pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blackhole-mock-server"
version = "0.5.0"
description = "ASGI mock server with truth for black-box scanner testing"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111,<1.0",
  "starlette>=0.37,<1.0",
  "pydantic>=2.7,<3.0",
  "jinja2>=3.1,<4.0",
  "pyyaml>=6.0,<7.0",
  "uvicorn>=0.30,<1.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["blackhole*"]
